+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # postcss-relative-rem Change Log
 
+## 1.0.0 - 2024-06-18
+
+### Fixed
+
+- Fix handling of negative values (eg, `-1rem`).
+- Fix handling of default values inside variables (eg, `var(--example-var, 2rem)`).
+
+### Changed
+
+- Convert project to TypeScript.
+- Add test coverage.
+
 ## 0.1.0 - 2022-02-15
 
 ### Added
+
 - Initial release.